Terminando modulo project-members

+++ app/Http/Controllers/ProjectController.php
@@ -39,7 +39,7 @@ class ProjectController extends Controller
 	 * @return Response
 	 */
 	public function index(){
-		return $this->service->index();
+		return $this->service->findWhere(['owner_id' => \Authorizer::getResourceOwnerId()]);
 	}
 	
 	/**
@@ -58,6 +58,11 @@ class ProjectController extends Controller
 	 * @return Response
 	 */
 	public function update(Request $request, $id){
+		
+		if ($this->checkProjectOwer($id) == false){
+			return ['error' => 'Access Forbidden'];
+		}
+		
 		return $this->service->update($request->all(), $id);
 	}
 
@@ -67,6 +72,11 @@ class ProjectController extends Controller
 	 * @return string|Exception
 	 */
 	public function destroy($id){
+		
+		if ($this->checkProjectOwer($id) == false){
+			return ['error' => 'Access Forbidden'];
+		}
+		
 		return $this->service->destroy($id);
 	}
 	
@@ -82,6 +92,11 @@ class ProjectController extends Controller
 		//if($this->repository->isOwner($id, $userId) == false){
 		//	return ['success' => false];
 		//}
+		
+		if ($this->checkProjectPermissions($id) == false){
+			return ['error' => 'Access Forbidden'];	
+		}
+		
 		return $this->service->show($id);
 	}
 	
@@ -108,5 +123,27 @@ class ProjectController extends Controller
 	
 	public function members($project_id){
 		return $this->service->isMember($project_id);
+	}
+	
+	//Metodo para verificar a autorização
+	private function checkProjectOwer($projectId){
+		//Pegar o id do user
+		$userId = \Authorizer::getResourceOwnerId();
+		
+		return $this->repository->isOwner($projectId, $userId);
+	}
+	
+	private function checkProjectMember($projectId){
+		//Pegar o id do user
+		$userId = \Authorizer::getResourceOwnerId();
+	
+		return $this->repository->hasMember($projectId, $userId);
+	}
+	
+	private function checkProjectPermissions($projectId){
+		if($this->checkProjectOwer($projectId) or $this->checkProjectMember($projectId)){
+			return true;
+		}
+			return false;
 	}
 }
